1. user UI implementation2. added customer Login and registration in main window3. deleted remove button from customer and station window4. added icon to all program windows

+++ BL/BL/BL.cs
@@ -408,7 +408,7 @@ namespace BL
             int toCount = cstmr.To.Count(prc => prc.Status == ParcelStatus.Delivered);
             return new CustomerInList
             {
-                Id = cstmr.Id,
+                Id = (int)cstmr.Id,
                 Name = cstmr.Name,
                 Phone = cstmr.Phone,
                 SentCount = fromCount,
